pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "score-core"
version = "0.1.3"
description = "Core library for the Score portable AI knowledge format"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "MultipleWorks", email = "hello@multipleworks.com.hk"},
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["ai", "llm", "governance", "skills", "score"]
dependencies = [
    "pydantic>=2.0",
    "ruamel.yaml>=0.18",
    "python-frontmatter>=1.0",
    "click>=8.0",
    "eval_type_backport>=0.2.0; python_version < '3.10'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.urls]
Homepage = "https://github.com/multipleworks/score"
Documentation = "https://github.com/multipleworks/score/blob/main/score_spec.md"
Repository = "https://github.com/multipleworks/score"

[project.scripts]
score = "score.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/score"]

[tool.hatch.build.targets.sdist]
include = [
    "src/score",
    "README.md",
    "LICENCE",
    "CHANGELOG.md",
    "docs",
]
